requirements.txt

langchain
langchain-community
langchain-core
langchain-ollama
langchain-chroma
sentence-transformers
chromadb
pdfminer.six
flask
flask-cors
python-dotenv
requests
accelerate
safetensors
sentencepiece
langdetect
